+++ spec/lib/govuk_conditional_content_item_loader_test_helpers_spec.rb
@@ -53,4 +53,26 @@ RSpec.describe GovukConditionalContentItemLoaderTestHelpers do
       expect { loader.load }.to raise_error(GdsApi::Base::ItemNotFound)
     end
   end
+
+  describe ".stub_conditional_content_loader_isnt_available" do
+    before do
+      stub_conditional_content_loader_isnt_available
+    end
+
+    it "returns an error for the base path" do
+      loader = GovukConditionalContentItemLoader.new(request:)
+      expect { loader.load }.to raise_error(GdsApi::HTTPUnavailable)
+    end
+  end
+
+  describe ".stub_conditional_content_loader_has_gone_item" do
+    before do
+      stub_conditional_content_loader_has_gone_item(base_path)
+    end
+
+    it "returns an error for the base path" do
+      loader = GovukConditionalContentItemLoader.new(request:)
+      expect { loader.load }.to raise_error(GdsApi::HTTPGone)
+    end
+  end
 end