+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.1
+
+* Add `stub_conditional_content_loader_isnt_available` test helper.
+* Add `stub_conditional_content_loader_has_gone_item` test helper.
+
 ## 2.0.0
 
 * BREAKING: replaces `stub_conditional_loader_returns_content_item` test helper with `stub_conditional_loader_returns_content_item_for_path` to match requests to a specific request path, for testing in frontend applications.